feat: 프로필 정보 입력 컴포넌트 분리(#40)
